Blender-related tests in test_blender.py
